requirements.txt
click==8.0.1
Flask==2.0.1
gensim==4.1.2
gunicorn==20.1.0
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mecab-python3==1.0.4
numpy==1.21.2
pandas==1.3.3
python-dateutil==2.8.2
pytz==2021.1
scipy==1.7.1
six==1.16.0
smart-open==5.2.1
typing-extensions==3.10.0.2
unidic-lite==1.0.8
Werkzeug==2.0.1
zipp==3.5.0
